+++ package/Codegen/Model.cs
@@ -133,6 +133,12 @@ namespace Codegen
         public string TS { get; set; }
     }
 
+    public class Command
+    {
+        public string Name { get; set; }
+        public string TSArgTypes { get; set; }
+    }
+
     public partial class VersionHeader
     {
         public Version Version { get; set; }