+++ package/Codegen/TSProps.cs
@@ -18,7 +18,7 @@ namespace Codegen
     /// Class to produce the template output
     /// </summary>
     
-    #line 1 "C:\Users\asklar\source\repos\react-native-xaml\package\Codegen\TSProps.tt"
+    #line 1 "F:\react-native-xaml\package\Codegen\TSProps.tt"
     [global::System.CodeDom.Compiler.GeneratedCodeAttribute("Microsoft.VisualStudio.TextTemplating", "16.0.0.0")]
     public partial class TSProps : TSPropsBase
     {
@@ -42,16 +42,18 @@ export type GridLength = number | '*' | 'auto' | `${number}*`;
 
 export type CornerRadius = number | { topLeft?: number, topRight?: number, bottomLeft?: number, bottomRight?: number};
 
+export type Point = { x: number, y: number };
+
 ");
             
-            #line 20 "C:\Users\asklar\source\repos\react-native-xaml\package\Codegen\TSProps.tt"
+            #line 22 "F:\react-native-xaml\package\Codegen\TSProps.tt"
  var derived = Util.GetDerivedTypes(Types);
 foreach (var type in Types) { 
             
             #line default
             #line hidden
             
-            #line 22 "C:\Users\asklar\source\repos\react-native-xaml\package\Codegen\TSProps.tt"
+            #line 24 "F:\react-native-xaml\package\Codegen\TSProps.tt"
  var ns = Util.GetTSNamespace(type);
 if (ns != "") { 
             
@@ -59,35 +61,35 @@ if (ns != "") {
             #line hidden
             this.Write("export namespace Native");
             
-            #line 24 "C:\Users\asklar\source\repos\react-native-xaml\package\Codegen\TSProps.tt"
+            #line 26 "F:\react-native-xaml\package\Codegen\TSProps.tt"
             this.Write(this.ToStringHelper.ToStringWithCulture(ns));
             
             #line default
             #line hidden
             this.Write(" {\r\n");
             
-            #line 25 "C:\Users\asklar\source\repos\react-native-xaml\package\Codegen\TSProps.tt"
+            #line 27 "F:\react-native-xaml\package\Codegen\TSProps.tt"
  } 
             
             #line default
             #line hidden
             this.Write("export interface Native");
             
-            #line 26 "C:\Users\asklar\source\repos\react-native-xaml\package\Codegen\TSProps.tt"
+            #line 28 "F:\react-native-xaml\package\Codegen\TSProps.tt"
             this.Write(this.ToStringHelper.ToStringWithCulture(Util.ToJsName(type)));
             
             #line default
             #line hidden
             this.Write("Props extends ");
             
-            #line 26 "C:\Users\asklar\source\repos\react-native-xaml\package\Codegen\TSProps.tt"
+            #line 28 "F:\react-native-xaml\package\Codegen\TSProps.tt"
             this.Write(this.ToStringHelper.ToStringWithCulture(Util.GetBaseClassProps(type)));
             
             #line default
             #line hidden
             this.Write(" {\r\n");
             
-            #line 27 "C:\Users\asklar\source\repos\react-native-xaml\package\Codegen\TSProps.tt"
+            #line 29 "F:\react-native-xaml\package\Codegen\TSProps.tt"
  var typeProp = Util.GetJsTypeProperty(type, derived);
 if (typeProp != "") { 
             
@@ -95,14 +97,14 @@ if (typeProp != "") {
             #line hidden
             this.Write("  type: ");
             
-            #line 29 "C:\Users\asklar\source\repos\react-native-xaml\package\Codegen\TSProps.tt"
+            #line 31 "F:\react-native-xaml\package\Codegen\TSProps.tt"
             this.Write(this.ToStringHelper.ToStringWithCulture(typeProp));
             
             #line default
             #line hidden
             this.Write(";\r\n");
             
-            #line 30 "C:\Users\asklar\source\repos\react-native-xaml\package\Codegen\TSProps.tt"
+            #line 32 "F:\react-native-xaml\package\Codegen\TSProps.tt"
  }
 foreach (var prop in type.GetProperties().Where(p => Util.ShouldEmitPropertyMetadata(p))) { 
             
@@ -110,21 +112,21 @@ foreach (var prop in type.GetProperties().Where(p => Util.ShouldEmitPropertyMeta
             #line hidden
             this.Write("  ");
             
-            #line 32 "C:\Users\asklar\source\repos\react-native-xaml\package\Codegen\TSProps.tt"
+            #line 34 "F:\react-native-xaml\package\Codegen\TSProps.tt"
             this.Write(this.ToStringHelper.ToStringWithCulture(Util.ToJsName(prop)));
             
             #line default
             #line hidden
             this.Write("?: ");
             
-            #line 32 "C:\Users\asklar\source\repos\react-native-xaml\package\Codegen\TSProps.tt"
+            #line 34 "F:\react-native-xaml\package\Codegen\TSProps.tt"
             this.Write(this.ToStringHelper.ToStringWithCulture(Util.GetTypeScriptType(prop)));
             
             #line default
             #line hidden
             this.Write(";\r\n");
             
-            #line 33 "C:\Users\asklar\source\repos\react-native-xaml\package\Codegen\TSProps.tt"
+            #line 35 "F:\react-native-xaml\package\Codegen\TSProps.tt"
  } 
 foreach (var prop in FakeProps.Where(p => type == p.DeclaringType)) { 
             
@@ -132,21 +134,21 @@ foreach (var prop in FakeProps.Where(p => type == p.DeclaringType)) {
             #line hidden
             this.Write("  ");
             
-            #line 35 "C:\Users\asklar\source\repos\react-native-xaml\package\Codegen\TSProps.tt"
+            #line 37 "F:\react-native-xaml\package\Codegen\TSProps.tt"
             this.Write(this.ToStringHelper.ToStringWithCulture(Util.ToJsName(prop)));
             
             #line default
             #line hidden
             this.Write("?: ");
             
-            #line 35 "C:\Users\asklar\source\repos\react-native-xaml\package\Codegen\TSProps.tt"
+            #line 37 "F:\react-native-xaml\package\Codegen\TSProps.tt"
             this.Write(this.ToStringHelper.ToStringWithCulture(Util.GetTypeScriptType(prop)));
             
             #line default
             #line hidden
             this.Write("; // synthetic property\r\n");
             
-            #line 36 "C:\Users\asklar\source\repos\react-native-xaml\package\Codegen\TSProps.tt"
+            #line 38 "F:\react-native-xaml\package\Codegen\TSProps.tt"
  } 
 foreach (var prop in SyntheticProps.Where(p => type == p.DeclaringType)) { 
             
@@ -154,28 +156,28 @@ foreach (var prop in SyntheticProps.Where(p => type == p.DeclaringType)) {
             #line hidden
             this.Write("  /**\r\n  * ");
             
-            #line 39 "C:\Users\asklar\source\repos\react-native-xaml\package\Codegen\TSProps.tt"
+            #line 41 "F:\react-native-xaml\package\Codegen\TSProps.tt"
             this.Write(this.ToStringHelper.ToStringWithCulture(prop.Comment));
             
             #line default
             #line hidden
             this.Write("\r\n  */\r\n  ");
             
-            #line 41 "C:\Users\asklar\source\repos\react-native-xaml\package\Codegen\TSProps.tt"
+            #line 43 "F:\react-native-xaml\package\Codegen\TSProps.tt"
             this.Write(this.ToStringHelper.ToStringWithCulture(Util.ToJsName(prop.Name)));
             
             #line default
             #line hidden
             this.Write("?: ");
             
-            #line 41 "C:\Users\asklar\source\repos\react-native-xaml\package\Codegen\TSProps.tt"
+            #line 43 "F:\react-native-xaml\package\Codegen\TSProps.tt"
             this.Write(this.ToStringHelper.ToStringWithCulture(Util.GetTypeScriptType(prop)));
             
             #line default
             #line hidden
             this.Write("; // synthetic property\r\n");
             
-            #line 42 "C:\Users\asklar\source\repos\react-native-xaml\package\Codegen\TSProps.tt"
+            #line 44 "F:\react-native-xaml\package\Codegen\TSProps.tt"
  } 
   foreach (var attachedDP in Util.AttachedProperties.Where(p => Properties.Any(sp => sp.Property == p && sp.DeclaringType == type))) {
             
@@ -183,21 +185,21 @@ foreach (var prop in SyntheticProps.Where(p => type == p.DeclaringType)) {
             #line hidden
             this.Write("  ");
             
-            #line 44 "C:\Users\asklar\source\repos\react-native-xaml\package\Codegen\TSProps.tt"
+            #line 46 "F:\react-native-xaml\package\Codegen\TSProps.tt"
             this.Write(this.ToStringHelper.ToStringWithCulture(Util.ToJsName(attachedDP)));
             
             #line default
             #line hidden
             this.Write("?: ");
             
-            #line 44 "C:\Users\asklar\source\repos\react-native-xaml\package\Codegen\TSProps.tt"
+            #line 46 "F:\react-native-xaml\package\Codegen\TSProps.tt"
             this.Write(this.ToStringHelper.ToStringWithCulture(Util.GetTypeScriptType(attachedDP)));
             
             #line default
             #line hidden
             this.Write("; // attached property\r\n");
             
-            #line 45 "C:\Users\asklar\source\repos\react-native-xaml\package\Codegen\TSProps.tt"
+            #line 47 "F:\react-native-xaml\package\Codegen\TSProps.tt"
  }
   foreach (var evt in type.GetEvents()) { 
             
@@ -205,14 +207,14 @@ foreach (var prop in SyntheticProps.Where(p => type == p.DeclaringType)) {
             #line hidden
             this.Write("  on");
             
-            #line 47 "C:\Users\asklar\source\repos\react-native-xaml\package\Codegen\TSProps.tt"
+            #line 49 "F:\react-native-xaml\package\Codegen\TSProps.tt"
             this.Write(this.ToStringHelper.ToStringWithCulture(evt.GetName()));
             
             #line default
             #line hidden
             this.Write("?: (event: NativeSyntheticEvent<undefined>) => void;\r\n");
             
-            #line 48 "C:\Users\asklar\source\repos\react-native-xaml\package\Codegen\TSProps.tt"
+            #line 50 "F:\react-native-xaml\package\Codegen\TSProps.tt"
  } 
   foreach (var evt in SyntheticEvents.Where(e => type == e.DeclaringType)) { 
             
@@ -220,48 +222,48 @@ foreach (var prop in SyntheticProps.Where(p => type == p.DeclaringType)) {
             #line hidden
             this.Write("  on");
             
-            #line 50 "C:\Users\asklar\source\repos\react-native-xaml\package\Codegen\TSProps.tt"
+            #line 52 "F:\react-native-xaml\package\Codegen\TSProps.tt"
             this.Write(this.ToStringHelper.ToStringWithCulture(evt.Name));
             
             #line default
             #line hidden
             this.Write("?: (event: NativeSyntheticEvent<");
             
-            #line 50 "C:\Users\asklar\source\repos\react-native-xaml\package\Codegen\TSProps.tt"
+            #line 52 "F:\react-native-xaml\package\Codegen\TSProps.tt"
             this.Write(this.ToStringHelper.ToStringWithCulture(Util.GetTypeScriptType(evt)));
             
             #line default
             #line hidden
             this.Write(">) => void;\r\n");
             
-            #line 51 "C:\Users\asklar\source\repos\react-native-xaml\package\Codegen\TSProps.tt"
+            #line 53 "F:\react-native-xaml\package\Codegen\TSProps.tt"
  } 
             
             #line default
             #line hidden
             this.Write("}\r\n");
             
-            #line 53 "C:\Users\asklar\source\repos\react-native-xaml\package\Codegen\TSProps.tt"
+            #line 55 "F:\react-native-xaml\package\Codegen\TSProps.tt"
  if (ns != "") { 
             
             #line default
             #line hidden
             this.Write("}\r\n");
             
-            #line 55 "C:\Users\asklar\source\repos\react-native-xaml\package\Codegen\TSProps.tt"
+            #line 57 "F:\react-native-xaml\package\Codegen\TSProps.tt"
  } 
             
             #line default
             #line hidden
             
-            #line 56 "C:\Users\asklar\source\repos\react-native-xaml\package\Codegen\TSProps.tt"
+            #line 58 "F:\react-native-xaml\package\Codegen\TSProps.tt"
  } 
             
             #line default
             #line hidden
             this.Write("\r\nexport type XamlControlProps =");
             
-            #line 58 "C:\Users\asklar\source\repos\react-native-xaml\package\Codegen\TSProps.tt"
+            #line 60 "F:\react-native-xaml\package\Codegen\TSProps.tt"
  var first = true; foreach (var type in Types) {
 	if (Util.HasCtor(type) || !type.IsSealed) {
 		if (!first) 
@@ -270,7 +272,7 @@ foreach (var prop in SyntheticProps.Where(p => type == p.DeclaringType)) {
             #line hidden
             this.Write(" | ");
             
-            #line 60 "C:\Users\asklar\source\repos\react-native-xaml\package\Codegen\TSProps.tt"
+            #line 62 "F:\react-native-xaml\package\Codegen\TSProps.tt"
  ;
 		
 		first = false;
@@ -280,14 +282,14 @@ foreach (var prop in SyntheticProps.Where(p => type == p.DeclaringType)) {
             #line hidden
             this.Write(" ");
             
-            #line 63 "C:\Users\asklar\source\repos\react-native-xaml\package\Codegen\TSProps.tt"
+            #line 65 "F:\react-native-xaml\package\Codegen\TSProps.tt"
             this.Write(this.ToStringHelper.ToStringWithCulture(Util.GetNativePropsName(type)));
             
             #line default
             #line hidden
             this.Write("\r\n\t");
             
-            #line 64 "C:\Users\asklar\source\repos\react-native-xaml\package\Codegen\TSProps.tt"
+            #line 66 "F:\react-native-xaml\package\Codegen\TSProps.tt"
  }
 	}