+++ package/windows/ReactNativeXaml/Codegen/Version.g.h
@@ -7,4 +7,4 @@ THIS FILE WAS AUTOMATICALLY GENERATED, DO NOT MODIFY MANUALLY
 #define VERSION_MAJOR               0
 #define VERSION_MINOR               0
 #define VERSION_REVISION            0
-#define VERSION_BUILD               52
+#define VERSION_BUILD               53