+++ package/windows/ReactNativeXaml/XamlObject.cpp
@@ -9,6 +9,7 @@
 
 #include <UI.Text.h>
 #include "Codegen/EventArgsTypeProperties.g.h"
+#include "XamlViewManager.h"
 
 namespace jsi = facebook::jsi;
 
@@ -253,3 +254,16 @@ std::vector<jsi::PropNameID> XamlObject::getPropertyNames(jsi::Runtime& rt) noex
   return pnames;
 }
 
+auto ReadPoint(const JSValueObject& o) {
+  const auto& x = o["x"];
+  
+  return winrt::Windows::Foundation::Point{ o["x"].AsSingle(), o["y"].AsSingle() };
+}
+
+void ShowAtCommand(xaml::FrameworkElement fe, const winrt::Microsoft::ReactNative::JSValueArray& args, const XamlMetadata& xaml) noexcept {
+  const auto& point = ReadPoint(args[0].GetObjectProperty("point").AsObject());
+  if (auto menuFlyout = Unwrap<MenuFlyout>(fe)) {
+    auto target = xaml.GetFlyoutTarget(menuFlyout);
+    menuFlyout.ShowAt(target, point);
+  }
+}